TOP VIEW 2D 게임에서의 오브젝트와 텍스쳐간 상대적 위치

+++ 39강 - Animation(4)/CAnimation.cpp
@@ -82,7 +82,7 @@ void CAnimation::Create(CTexture* _pTex, Vec2 _vLT, Vec2 _vSliceSize, Vec2 _vSte
 	{
 		frm.fDuration = _fDuration;
 		frm.vSlice = _vSliceSize;
-		frm.vLT = _vLT + _vStep * i;
+		frm.vLT = _vLT + _vStep * (float)i;
 
 		m_vecFrm.push_back(frm);
 	}
